+++ src/main/java/org/nrg/containers/services/DockerService.java
@@ -4,7 +4,8 @@ import org.nrg.containers.exceptions.DockerServerException;
 import org.nrg.containers.exceptions.NoServerPrefException;
 import org.nrg.containers.exceptions.NotUniqueException;
 import org.nrg.containers.model.command.auto.Command;
-import org.nrg.containers.model.dockerhub.DockerHub;
+import org.nrg.containers.model.dockerhub.DockerHubBase.DockerHub;
+import org.nrg.containers.model.dockerhub.DockerHubBase.DockerHubWithPing;
 import org.nrg.containers.model.image.docker.DockerImage;
 import org.nrg.containers.model.image.docker.DockerImageAndCommandSummary;
 import org.nrg.containers.model.server.docker.DockerServerBase;
@@ -16,11 +17,11 @@ import org.nrg.prefs.exceptions.InvalidPreferenceName;
 import java.util.List;
 
 public interface DockerService {
-    List<DockerHub> getHubs();
-    DockerHub getHub(long id) throws NotFoundException;
-    DockerHub getHub(String name) throws NotFoundException, NotUniqueException;
-    DockerHub createHub(DockerHub hub);
-    DockerHub createHubAndSetDefault(DockerHub hub, String username, String reason);
+    List<DockerHubWithPing> getHubs();
+    DockerHubWithPing getHub(long id) throws NotFoundException;
+    DockerHubWithPing getHub(String name) throws NotFoundException, NotUniqueException;
+    DockerHubWithPing createHub(DockerHub hub);
+    DockerHubWithPing createHubAndSetDefault(DockerHub hub, String username, String reason);
     void updateHub(DockerHub hub);
     void updateHubAndSetDefault(DockerHub hub, String username, String reason);
     void setDefaultHub(long id, String username, String reason);